+++ resources/views/tailwind/includes/filters.blade.php
@@ -8,7 +8,7 @@
         <div>
             <button
                 type="button"
-                class="inline-flex justify-center w-full rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:border-indigo-300 focus:shadow-outline-indigo"
+                class="inline-flex justify-center w-full rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-sm font-medium text-gray-700 hover:bg-gray-50 focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50 dark:bg-gray-700 dark:text-white dark:border-gray-600 dark:hover:bg-gray-600"
                 id="filters-menu"
                 x-on:click="open = !open"
                 aria-haspopup="true"
@@ -40,7 +40,7 @@
             x-transition:leave="transition ease-in duration-75"
             x-transition:leave-start="transform opacity-100 scale-100"
             x-transition:leave-end="transform opacity-0 scale-95"
-            class="origin-top-right absolute right-0 mt-2 w-full md:w-56 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 divide-y divide-gray-100 focus:outline-none z-50"
+            class="origin-top-right absolute right-0 mt-2 w-full md:w-56 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 divide-y divide-gray-100 focus:outline-none z-50 dark:bg-gray-700 dark:text-white dark:divide-gray-600"
             role="menu"
             aria-orientation="vertical"
             aria-labelledby="filters-menu"
@@ -52,7 +52,7 @@
                     <div class="py-1" role="none">
                         <div class="block px-4 py-2 text-sm text-gray-700" role="menuitem">
                             <label for="filter-{{ $key }}"
-                                   class="block text-sm font-medium leading-5 text-gray-700">
+                                   class="block text-sm font-medium leading-5 text-gray-700 dark:text-white">
                                 {{ $filter->name() }}
                             </label>
 
@@ -68,12 +68,12 @@
 
             @if (count($this->getFiltersWithoutSearch()))
                 <div class="py-1" role="none">
-                    <div class="block px-4 py-2 text-sm text-gray-700" role="menuitem">
+                    <div class="block px-4 py-2 text-sm text-gray-700 dark:text-white" role="menuitem">
                         <button
                             wire:click.prevent="resetFilters"
                             x-on:click="open = false"
                             type="button"
-                            class="w-full inline-flex items-center justify-center px-3 py-2 border border-gray-300 shadow-sm text-sm leading-4 font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
+                            class="w-full inline-flex items-center justify-center px-3 py-2 border border-gray-300 shadow-sm text-sm leading-4 font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50 dark:bg-gray-800 dark:border-gray-600 dark:text-white dark:hover:border-gray-500"
                         >
                             @lang('Clear')
                         </button>