+++ resources/views/tailwind/includes/pagination.blade.php
@@ -3,7 +3,7 @@
         @if ($paginationEnabled && $rows->lastPage() > 1)
             {{ $rows->links() }}
         @else
-            <p class="text-sm text-gray-700 leading-5">
+            <p class="text-sm text-gray-700 leading-5 dark:text-white">
                 @lang('Showing')
                 <span class="font-medium">{{ $rows->count() }}</span>
                 @lang('results')