+++ resources/views/tailwind/includes/sorting-pills.blade.php
@@ -1,7 +1,7 @@
 <div>
     @if ($showSorting && count($sorts))
         <div class="mb-4 p-6 md:p-0">
-            <small class="text-gray-700">@lang('Applied Sorting'):</small>
+            <small class="text-gray-700 dark:text-white">@lang('Applied Sorting'):</small>
 
             @foreach($sorts as $col => $dir)
                 <span